refactor: move server args into own file
